refactor: rename tui util module to utils
